refactor(astro-service): extract electional workflows
